+++ CHANGES.md
@@ -79,6 +79,7 @@ Changes in 12.0
 - Remove assets/PortalNotesFiles/ if only contains README in Update.fnc.php
 - Remove Cookie vars from URL, may be included in `$_REQUEST` in PreparePHP_SELF.fnc.php
 - Deprecate `_ReindexResults()` function in ListOutput.fnc.php
+- Search List: Remove Relevance column, do not sort results in ListOutput.fnc.php
 
 Changes in 11.8.6
 -----------------